+++ ql/FiniteDifferences/forwardeuler.hpp
@@ -56,14 +56,14 @@ namespace QuantLib {
             Operator& operator=(const Operator&);
 
             // inspectors
-            unsigned int size(); 
+            size_t size(); 
 
             // modifiers
             void setTime(Time t); 
 
             // operator interface
             arrayType applyTo(const arrayType&);
-            static Operator identity(unsigned int size);
+            static Operator identity(size_t size);
 
             // operator algebra
             Operator operator*(double, const Operator&);
