+++ ql/Math/multivariateaccumulator.hpp
@@ -51,11 +51,11 @@ namespace QuantLib {
         class MultivariateAccumulator {
           public:
             MultivariateAccumulator();
-            MultivariateAccumulator(unsigned int size);
+            MultivariateAccumulator(size_t size);
             //! \name Inspectors
             //@{
             //! size of each sample
-            unsigned int size() const;
+            size_t size() const;
             //! number of samples collected
             double samples() const;
             //! returns the mean as an Array
@@ -89,7 +89,7 @@ namespace QuantLib {
             void reset();
             //@}
           private:
-            unsigned int size_;
+            size_t size_;
             double sampleNumber_;
             double sampleWeight_;
             Array sum_;
@@ -98,7 +98,7 @@ namespace QuantLib {
 
 
         // inline definitions
-        inline unsigned int MultivariateAccumulator::size() const {
+        inline size_t MultivariateAccumulator::size() const {
             return size_;
         }
