+++ ql/MonteCarlo/europeanpathpricer.cpp
@@ -52,12 +52,12 @@ namespace QuantLib {
           antitheticVariance) {}
 
         double EuropeanPathPricer::operator()(const Path & path) const {
-            unsigned int n = path.size();
+            size_t n = path.size();
             QL_REQUIRE(n>0,
                 "EuropeanPathPricer: the path cannot be empty");
 
             double log_drift = 0.0, log_random = 0.0;
-            for (unsigned int i = 0; i < n; i++) {
+            for (size_t i = 0; i < n; i++) {
                 log_drift += path.drift()[i];
                 log_random += path.diffusion()[i];
             }
